+++ lib/full_search/config.rb
@@ -4,7 +4,8 @@ module FullSearch
   class Config
     attr_accessor :auto_rebuild_schema, :stale_query_behavior, :lock_rebuilds,
       :default_async_reindex, :default_async_source_reindex,
-      :default_tokenizer, :auto_rebuild_on_stale_query, :min_like_prefix_length
+      :default_tokenizer, :auto_rebuild_on_stale_query, :min_like_prefix_length,
+      :dump_schema_virtual_tables
 
     def initialize
       @auto_rebuild_schema = false
@@ -15,6 +16,7 @@ module FullSearch
       @default_tokenizer = FullSearch::Constants::DEFAULT_TOKENIZER
       @auto_rebuild_on_stale_query = false
       @min_like_prefix_length = FullSearch::Constants::DEFAULT_MIN_LIKE_PREFIX_LENGTH
+      @dump_schema_virtual_tables = true
     end
   end