+++ lib/full_search/constants.rb
@@ -6,7 +6,7 @@ module FullSearch
     TWO_TYPO_MIN_LENGTH = 9
     DEFAULT_MIN_LIKE_PREFIX_LENGTH = 3
     REBUILDING_HASH = "__rebuilding__"
-    DEFAULT_TOKENIZER = "unicode61"
+    DEFAULT_TOKENIZER = "porter"
     MAX_EXACT_MATCH_BOOST_IDS = 100
   end
 end